Initial port from RapidApp::CatalystX::SimpleCAS::*
Ported the SimpleCAS code out of RapidApp core with some minor
tweaks/modifications to work in stand-alone structure. The main
controller appears to work via RapidApp, however, the TextTranscode
controller will need refactoring to tie to SimpleCAS rather than to
RapidApp. Also, this whole codebase is still pretty rough as-is and is
going to be refactored in general...