Converted TextTranscode into a Controller Role of SimpleCAS
The separate TextTranscode controller is now a Role of the main
SimpleCAS controller. The TextTranscode actions are provided at the same
relative URL paths as they were in the original implementation within
RapidApp which used to inject both controllers.

Assuming the SimpleCAS controler namespace is 'simplecas' (as in
RapidApp automatically) the TextTranscode action namespaces are still
at:

  /simplecas/texttranscode/transcode_html
  /simplecas/texttranscode/generate_mhtml_download

See the HtmlEditor in RapidApp as of version 0.99333 mor more details...

Note: the reason this role design is being used at the moment is simply
to keep the structure as close to the same as possible... this may
change later on...

Also note that as of this commit the generate_mhtml_download is
broken/buggy. It was already like that in RapidApp... This whole
TextTranscode stuff always has been an ambitious, experimental
feature...