Fix some warnings, but not from Email::Simple::Header
We call Email::MIME->new($data) to parse raw content in two pleases,
and depending on the content, this will produce warnings all the time
like:

 Use of uninitialized value in lc at <pth>/Email/Simple/Header.pm line 204.

And I couldn't figure out how to stop it. Setting no warnings 'uninitialized'
in the closest scope possible didn't work. I'm not sure why, but I've
wasted too much time on it for now...