Bypass broken CSS::Simple (busted on CPAN)
I hate doing this, but I just don't have the time to deal with it
properly... CSS::Simple was broken on CPAN for perl 5.12 and earlier
(version 3220 which is the latest as of this commit) which breaks us
and in turn breaks RapidApp. I reported the problem in RT#105857 but
there has been no response so far and I can't wait any more, so I've
created and switched to a local copy of the class as
Catalyst::Controller::SimpleCAS::CSS::Simple. I wish I had time to
have made a larger effort to follow up with the author, hopefully
I can try again, but for now, my module is fixed and deptree is
fixed...