Move sha1 functions back to Store, from Controller
Per discussion with Getty, the goal was to make sure a subclass
didn't change the hashing function and thereby break the CAS
when interchanged for other back-ends.

However, in the spirit of having the store be independent of the
controller, we agreed that it was better to have it in the Store
role where controllers could share it, and remove the dependence
on the controller instance.