Add Catalyst::Plugin::SimpleCAS, new sugar methods, and docs
Added a new plugin interface to use the module. This simply sets up a
controller automatically, which is trivial, but is being done for the
purposes of consistent usage/examples. RapidApp is changing to use this
new interface in its next CPAN release.

Also added sugar functions to make it easier to interact with the CAS in
backend scripts:

 * fetch
 * fetch_fh
 * add

Updated POD documentation, including a lot more explanation/detail of
the purpose and rationales of the module.